Tighten up compile warnings based to match other chromium sub-projects. Fix up the issues that the new warnings raised.
BUG=none
TEST=build remoting cleanly

Review URL: http://codereview.chromium.org/2801003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@49842 0039d316-1c4b-4281-b951-d872f2087c98

+++ remoting/base/multiple_array_input_stream_unittest.cc
@@ -10,8 +10,8 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace remoting {
 
-static int ReadFromInput(MultipleArrayInputStream* input,
-                         void* data, int size) {
+static size_t ReadFromInput(MultipleArrayInputStream* input,
+                         void* data, size_t size) {
   uint8* out = reinterpret_cast<uint8*>(data);
   int out_size = size;
