+++ remoting/host/client_connection.cc
@@ -16,7 +16,7 @@ namespace remoting {
 
 // Determine how many update streams we should count to find the size of
 // average update stream.
-static const int kAverageUpdateStream = 10;
+static const size_t kAverageUpdateStream = 10;
 
 ClientConnection::ClientConnection(MessageLoop* message_loop,
                                    ProtocolDecoder* decoder,
